# docs/phase2/29_API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: FlexSpace API
  version: "1.0.0"
  description: >
    FlexSpace Flexible Workspace Marketplace — Phase 2 API contract.
    Implements the decisions in 11_API_CONTRACTS.md. REST/JSON, versioned base path,
    JWT bearer auth, consistent error envelope, idempotency keys on mutating
    financial/booking endpoints. This is a contract, not an implementation —
    no server code exists yet.
servers:
  - url: https://api.flexspace.az/api/v1
    description: Production (placeholder domain)
  - url: https://staging-api.flexspace.az/api/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /spaces/{roomId}/reviews:
    # Missing from this file entirely until the /account/reviews pass —
    # ReviewsController.listForRoom() is real and already @Public(), the
    # path simply had never been documented. Added as doc-only completion,
    # not a new endpoint.
    get:
      tags: [Reviews]
      summary: Public reviews for a room (APPROVED moderation status only)
      security: []
      parameters:
        - { $ref: "#/components/parameters/RoomId" }
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Review" } }
...
  /reviews:
    post:
      tags: [Reviews]
      summary: Create a review (only for COMPLETED bookings — 18_SECURITY.md §18.6)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, rating]
              properties:
                bookingId: { type: string, format: uuid }
                rating: { type: integer, minimum: 1, maximum: 5 }
                text: { type: string }
                photos: { type: array, items: { type: string, format: uri } }
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
        "403":
          description: Booking not eligible for review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /providers:
    post:
      tags: [Provider]
      summary: Register as a provider (self-service — starts PENDING, 09_DOMAIN_MODEL.md §9.2)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderCreateRequest" }
      responses:
        "201":
          description: Provider created, verification_status = PENDING
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
...
  /admin/providers:
    get:
      tags: [Admin]
      summary: List providers by verification status
      parameters:
        - { name: verificationStatus, in: query, schema: { type: string } }
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Provider" } }
  /admin/providers/{providerId}/verify:
    post:
      tags: [Admin]
      summary: Approve/reject provider verification (24_ADMIN_ARCHITECTURE.md §24.2)
      parameters:
        - { name: providerId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [VERIFIED, REJECTED] }
                notes: { type: string }
      responses: { "200": { description: Verification decision recorded } }
...
  /admin/payouts:
    get:
      tags: [Admin]
      summary: Payout queue (14_PAYOUT_LEDGER.md §14.5), filterable by payee type
      description: >
        31_PARTNER_REFERRAL_ARCHITECTURE.md §31.6 — the payeeType filter is
        additive; omitting it returns both provider and partner payouts as
        before this addendum.
      parameters:
        - name: payeeType
          in: query
          required: false
          schema: { type: string, enum: [PROVIDER, PARTNER] }
      responses:
        "200":
          description: Payouts
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Payout" } }
  /admin/payouts/{payoutId}/mark-paid:
    post:
      tags: [Admin]
      summary: Mark a payout as paid with a bank reference
      parameters:
        - { name: payoutId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bankReference]
              properties:
                bankReference: { type: string }
      responses: { "200": { description: Payout marked paid } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "SLOT_UNAVAILABLE" }
            message: { type: string }
            details: { type: object }
...
    Money:
      type: object
      properties:
        amount: { type: integer, description: "minor units (qəpik)" }
        currency: { type: string, example: "AZN" }
...
    Review:
      # Matches ReviewEntity exactly (backend/src/modules/reviews/entities/
      # review.entity.ts) — a prior version of this schema only had
      # id/bookingId/rating/text/createdAt, missing roomId/customerUserId/
      # photos/providerReplyText/providerReplyAt/moderationStatus. Fixed as
      # doc-only drift found while adding /reviews/me and
      # /spaces/{roomId}/reviews to this file.
      type: object
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        customerUserId: { type: string, format: uuid }
        roomId: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string, nullable: true }
        photos: { type: array, items: { type: string, format: uri }, nullable: true }
        providerReplyText: { type: string, nullable: true }
        providerReplyAt: { type: string, format: date-time, nullable: true }
        moderationStatus: { type: string, enum: [PENDING, APPROVED, REJECTED] }
        createdAt: { type: string, format: date-time }
    Provider:
      type: object
      properties:
        id: { type: string, format: uuid }
        legalName: { type: string }
        displayName: { type: string }
        slug: { type: string }
        category: { type: string, nullable: true }
        verificationStatus: { type: string, enum: [PENDING, VERIFIED, REJECTED, SUSPENDED] }
        planTier: { type: string, enum: [FREE, STARTER, PRO, ENTERPRISE] }
        commissionPercentage: { type: number, nullable: true }
    ProviderCreateRequest:
      type: object
      required: [legalName, displayName]
      properties:
        legalName: { type: string }
        displayName: { type: string }
        category: { type: string }
...
    Payout:
      type: object
      properties:
        id: { type: string, format: uuid }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        amount: { $ref: "#/components/schemas/Money" }
        status: { type: string, enum: [PENDING, AVAILABLE, PROCESSING, PAID, FAILED, REVERSED] }
...